Don't generate inset cache if not used
Made slicing on sla faster
