+++ Pathfinding/PathFinder.cs
@@ -34,15 +34,10 @@ namespace MatterHackers.Pathfinding
 {
 	using System;
 	using System.IO;
-	using MatterHackers.Agg;
-	using MatterHackers.Agg.Image;
-	using MatterHackers.Agg.Transform;
 	using MatterHackers.Agg.VertexSource;
 	using QuadTree;
 	using Polygon = List<IntPoint>;
 	using Polygons = List<List<IntPoint>>;
-	using static System.Math;
-	using MatterHackers.Agg.ImageProcessing;
 
 	public class PathFinder
 	{
@@ -90,7 +85,7 @@ namespace MatterHackers.Pathfinding
 
 		public bool AllPathSegmentsAreInsideOutlines(Polygon pathThatIsInside, IntPoint startPoint, IntPoint endPoint, bool writeErrors = false, int layerIndex = -1)
 		{
-			if(this.OutlineData == null)
+			if (this.OutlineData == null)
 			{
 				return true;
 			}
@@ -156,7 +151,7 @@ namespace MatterHackers.Pathfinding
 
 				MovePointsInsideIfPossible(startPointIn, endPointIn, pathThatIsInside);
 
-				var cleanPath = pathThatIsInside.CleanClosedPolygon(InsetAmount/2);
+				var cleanPath = pathThatIsInside.CleanClosedPolygon(InsetAmount / 2);
 				pathThatIsInside.Clear();
 				pathThatIsInside.AddRange(cleanPath);
 
@@ -202,6 +197,30 @@ namespace MatterHackers.Pathfinding
 			return OutlineData.PointIsInside(intPoint) == QTPolygonsExtensions.InsideState.Inside;
 		}
 
+		private static void Remove0LengthSegments(IntPoint startPointIn, IntPoint endPointIn, Polygon pathThatIsInside)
+		{
+			if (pathThatIsInside.Count > 1)
+			{
+				IntPoint startPoint = startPointIn;
+				for (int i = 0; i < pathThatIsInside.Count - 1; i++)
+				{
+					IntPoint endPoint = pathThatIsInside[i];
+
+					if (endPoint == startPoint)
+					{
+						// and remove the end point (it is the same as the start point
+						pathThatIsInside.RemoveAt(i);
+						// don't advance past the start point
+						i--;
+					}
+					else
+					{
+						startPoint = endPoint;
+					}
+				}
+			}
+		}
+
 		private static void RemoveUTurnSegments(IntPoint startPointIn, IntPoint endPointIn, Polygon pathThatIsInside)
 		{
 			if (pathThatIsInside.Count > 1)
@@ -224,30 +243,6 @@ namespace MatterHackers.Pathfinding
 					else
 					{
 						startPoint = testPoint;
-					}
-				}
-			}
-		}
-
-		private static void Remove0LengthSegments(IntPoint startPointIn, IntPoint endPointIn, Polygon pathThatIsInside)
-		{
-			if (pathThatIsInside.Count > 1)
-			{
-				IntPoint startPoint = startPointIn;
-				for (int i = 0; i < pathThatIsInside.Count - 1; i++)
-				{
-					IntPoint endPoint = pathThatIsInside[i];
-
-					if (endPoint == startPoint)
-					{
-						// and remove the end point (it is the same as the start point
-						pathThatIsInside.RemoveAt(i);
-						// don't advance past the start point
-						i--;
-					}
-					else
-					{
-						startPoint = endPoint;
 					}
 				}
 			}
@@ -354,19 +349,34 @@ namespace MatterHackers.Pathfinding
 			return true;
 		}
 
-		private bool SegmentIsAllInside(IntPointNode lastAddedNode, IntPointNode crossingNode)
+		private void CutIntoSmallSegments(Polygon pathThatIsInside)
 		{
-			if (OutlineData.InsideCache == null)
+			var cutLength = InsetAmount;
+			// Make every segment be a maximum of cutLength long
+			if (pathThatIsInside.Count >= 2
+				&& InsetAmount > 0)
 			{
-				// check just the center point
-				return OutlineData.PointIsInside((lastAddedNode.Position + crossingNode.Position) / 2) == QTPolygonsExtensions.InsideState.Inside;
-			}
-			else
-			{
-				// check many points along the line
-				return OutlineData.PointIsInside((lastAddedNode.Position + crossingNode.Position) / 2) == QTPolygonsExtensions.InsideState.Inside
-					&& OutlineData.PointIsInside(lastAddedNode.Position + (crossingNode.Position - lastAddedNode.Position) / 4) == QTPolygonsExtensions.InsideState.Inside
-					&& OutlineData.PointIsInside(lastAddedNode.Position + (crossingNode.Position - lastAddedNode.Position) * 3 / 4) == QTPolygonsExtensions.InsideState.Inside;
+				var startIndex = PointIsInsideBoundary(pathThatIsInside[0]) ? 0 : 1;
+				IntPoint startPoint = pathThatIsInside[startIndex];
+
+				var endIndex = PointIsInsideBoundary(pathThatIsInside[pathThatIsInside.Count - 1]) ? pathThatIsInside.Count - 1 : pathThatIsInside.Count - 2;
+				var cutLengthSquared = cutLength * cutLength;
+				for (int i = startIndex; i <= pathThatIsInside.Count - 2; i++)
+				{
+					startPoint = pathThatIsInside[i];
+					IntPoint endPoint = pathThatIsInside[i + 1];
+
+					if ((endPoint - startPoint).LengthSquared() > cutLengthSquared)
+					{
+						int steps = (int)((endPoint - startPoint).Length() / cutLength);
+						for (int cut = 1; cut < steps; cut++)
+						{
+							IntPoint newPosition = startPoint + (endPoint - startPoint) * cut / steps;
+							pathThatIsInside.Insert(i + 1, newPosition);
+							i++;
+						}
+					}
+				}
 			}
 		}
 
@@ -403,7 +413,7 @@ namespace MatterHackers.Pathfinding
 			(int polyIndex, int pointIndex, IntPoint position) foundPolyPointPosition;
 			waypointAtPosition = null;
 			OutlineData.Polygons.MovePointInsideBoundary(position, out foundPolyPointPosition, OutlineData.EdgeQuadTrees, OutlineData.PointQuadTrees, OutlineData.PointIsInside);
-			if(foundPolyPointPosition.polyIndex == -1)
+			if (foundPolyPointPosition.polyIndex == -1)
 			{
 				// The point is already inside
 				var existingNode = OutlineData.Waypoints.FindNode(position, findNodeDist);
@@ -469,37 +479,6 @@ namespace MatterHackers.Pathfinding
 			}
 		}
 
-		private void CutIntoSmallSegments(Polygon pathThatIsInside)
-		{
-			var cutLength = InsetAmount;
-			// Make every segment be a maximum of cutLength long
-			if (pathThatIsInside.Count >= 2
-				&& InsetAmount > 0)
-			{
-				var startIndex = PointIsInsideBoundary(pathThatIsInside[0]) ? 0 : 1;
-				IntPoint startPoint = pathThatIsInside[startIndex];
-
-				var endIndex = PointIsInsideBoundary(pathThatIsInside[pathThatIsInside.Count-1]) ? pathThatIsInside.Count - 1 : pathThatIsInside.Count - 2;
-				var cutLengthSquared = cutLength * cutLength;
-				for (int i = startIndex; i <= pathThatIsInside.Count - 2; i++)
-				{
-					startPoint = pathThatIsInside[i];
-					IntPoint endPoint = pathThatIsInside[i+1];
-
-					if ((endPoint - startPoint).LengthSquared() > cutLengthSquared)
-					{
-						int steps = (int)((endPoint - startPoint).Length() / cutLength);
-						for (int cut = 1; cut < steps; cut++)
-						{
-							IntPoint newPosition = startPoint + (endPoint - startPoint) * cut / steps;
-							pathThatIsInside.Insert(i+1, newPosition);
-							i++;
-						}
-					}
-				}
-			}
-		}
-
 		private void MovePointsInsideIfPossible(IntPoint startPointIn, IntPoint endPointIn, Polygon pathThatIsInside)
 		{
 			// move every segment that can be inside the boundry to be within the boundry
@@ -512,11 +491,11 @@ namespace MatterHackers.Pathfinding
 					IntPoint endPoint = i < pathThatIsInside.Count - 2 ? pathThatIsInside[i + 1] : endPointIn;
 
 					IntPoint inPolyPosition;
-					if(OutlineData.MovePointAwayFromEdge(testPoint, InsetAmount, out inPolyPosition))
+					if (OutlineData.MovePointAwayFromEdge(testPoint, InsetAmount, out inPolyPosition))
 					{
 						// It moved so test if it is a good point
 						//if (OutlineData.Polygons.FindIntersection(startPoint, inPolyPosition, OutlineData.EdgeQuadTrees) != Intersection.Intersect
-							//&& OutlineData.Polygons.FindIntersection(inPolyPosition, endPoint, OutlineData.EdgeQuadTrees) != Intersection.Intersect)
+						//&& OutlineData.Polygons.FindIntersection(inPolyPosition, endPoint, OutlineData.EdgeQuadTrees) != Intersection.Intersect)
 						{
 							testPoint = inPolyPosition;
 							pathThatIsInside[i] = testPoint;
@@ -565,6 +544,22 @@ namespace MatterHackers.Pathfinding
 			}
 		}
 
+		private bool SegmentIsAllInside(IntPointNode lastAddedNode, IntPointNode crossingNode)
+		{
+			if (OutlineData.InsideCache == null)
+			{
+				// check just the center point
+				return OutlineData.PointIsInside((lastAddedNode.Position + crossingNode.Position) / 2) == QTPolygonsExtensions.InsideState.Inside;
+			}
+			else
+			{
+				// check many points along the line
+				return OutlineData.PointIsInside((lastAddedNode.Position + crossingNode.Position) / 2) == QTPolygonsExtensions.InsideState.Inside
+					&& OutlineData.PointIsInside(lastAddedNode.Position + (crossingNode.Position - lastAddedNode.Position) / 4) == QTPolygonsExtensions.InsideState.Inside
+					&& OutlineData.PointIsInside(lastAddedNode.Position + (crossingNode.Position - lastAddedNode.Position) * 3 / 4) == QTPolygonsExtensions.InsideState.Inside;
+			}
+		}
+
 		private bool ValidPoint(PathingData outlineData, IntPoint position)
 		{
 			(int polyIndex, int pointIndex, IntPoint position) movedPosition;
@@ -608,279 +603,6 @@ namespace MatterHackers.Pathfinding
 					sw.WriteLine($"// Length of this segment (start->end) {length}. Length of bad edge {edgeLength}");
 					sw.WriteLine($"// startOverride = new MSIntPoint({startPoint.X}, {startPoint.Y}); endOverride = new MSIntPoint({endPoint.X}, {endPoint.Y});");
 					sw.WriteLine($"TestSinglePathIsInside(polyPath, new IntPoint({startPoint.X}, {startPoint.Y}), new IntPoint({endPoint.X}, {endPoint.Y}));");
-				}
-			}
-		}
-	}
-
-	/// <summary>
-	/// This is to hold all the data that lets us switch between Boundry and Outline pathing.
-	/// </summary>
-	public class PathingData
-	{
-		private Affine polygonsToImageTransform;
-		private double unitsPerPixel;
-		private bool usingPathingCache;
-		IntRect polygonBounds;
-
-		internal PathingData(Polygons polygons, double unitsPerPixel, bool usingPathingCache)
-		{
-			this.usingPathingCache = usingPathingCache;
-
-			Polygons = polygons;
-			polygonBounds = Polygons.GetBounds();
-			SetGoodUnitsPerPixel(unitsPerPixel);
-
-			EdgeQuadTrees = Polygons.GetEdgeQuadTrees();
-			PointQuadTrees = Polygons.GetPointQuadTrees();
-
-			foreach (var polygon in Polygons)
-			{
-				Waypoints.AddPolygon(polygon);
-			}
-
-			RemovePointList = new WayPointsToRemove(Waypoints);
-
-			GenerateIsideCache();
-		}
-
-		const int maxImageSize = 4096;
-		private void SetGoodUnitsPerPixel(double unitsPerPixel)
-		{
-			unitsPerPixel = Max(unitsPerPixel, 1);
-			if (polygonBounds.Width() / unitsPerPixel > maxImageSize)
-			{
-				unitsPerPixel = Max(1, polygonBounds.Width() / maxImageSize);
-			}
-			if (polygonBounds.Height() / unitsPerPixel > maxImageSize)
-			{
-				unitsPerPixel = Max(1, polygonBounds.Height() / maxImageSize);
-			}
-			if (polygonBounds.Width() / unitsPerPixel < 32)
-			{
-				unitsPerPixel = polygonBounds.Width() / 32;
-			}
-			if (polygonBounds.Height() / unitsPerPixel > maxImageSize)
-			{
-				unitsPerPixel = polygonBounds.Height() / 32;
-			}
-
-			this.unitsPerPixel = Max(1, unitsPerPixel);
-		}
-
-		public List<QuadTree<int>> EdgeQuadTrees { get; }
-		public ImageBuffer InsideCache { get; private set; }
-		public ImageBuffer InsetMap { get; private set; }
-		public List<QuadTree<int>> PointQuadTrees { get; }
-		public Polygons Polygons { get; }
-		public WayPointsToRemove RemovePointList { get; }
-		public IntPointPathNetwork Waypoints { get; } = new IntPointPathNetwork();
-
-		public static VertexStorage CreatePathStorage(List<List<IntPoint>> polygons)
-		{
-			VertexStorage output = new VertexStorage();
-
-			foreach (List<IntPoint> polygon in polygons)
-			{
-				bool first = true;
-				foreach (IntPoint point in polygon)
-				{
-					if (first)
-					{
-						output.Add(point.X, point.Y, ShapePath.FlagsAndCommand.CommandMoveTo);
-						first = false;
-					}
-					else
-					{
-						output.Add(point.X, point.Y, ShapePath.FlagsAndCommand.CommandLineTo);
-					}
-				}
-
-				output.ClosePolygon();
-			}
-
-			return output;
-		}
-
-		public bool MovePointAwayFromEdge(IntPoint testPoint, long distance, out IntPoint result)
-		{
-			int distanceInPixels = (int)Round(distance / unitsPerPixel);
-			result = testPoint;
-			bool movedPoint = false;
-
-			for (int i = 0; i < distanceInPixels + distanceInPixels/2; i++)
-			{
-				// check each direction to see if we can increase our InsetMap value
-				double x = result.X;
-				double y = result.Y;
-				polygonsToImageTransform.transform(ref x, ref y);
-				int xi = (int)Round(x);
-				int yi = (int)Round(y);
-
-				int current = GetInsetMapValue(xi, yi);
-				if(current == 255)
-				{
-					// we've made it all the way inside
-				}
-
-				var offset = new IntPoint();
-				movedPoint |= CheckInsetPixel(current, xi, - 1, yi, + 0, ref offset);
-				movedPoint |= CheckInsetPixel(current, xi, - 1, yi, - 1, ref offset);
-				movedPoint |= CheckInsetPixel(current, xi, + 0, yi, - 1, ref offset);
-				movedPoint |= CheckInsetPixel(current, xi, + 1, yi, - 1, ref offset);
-				movedPoint |= CheckInsetPixel(current, xi, + 1, yi, + 0, ref offset);
-				movedPoint |= CheckInsetPixel(current, xi, + 1, yi, + 1, ref offset);
-				movedPoint |= CheckInsetPixel(current, xi, + 0, yi, + 1, ref offset);
-				movedPoint |= CheckInsetPixel(current, xi, - 1, yi, + 1, ref offset);
-
-				if (offset.X < 0) x -= 1; else if (offset.X > 0) x += 1;
-				if (offset.Y < 0) y -= 1; else if (offset.Y > 0) y += 1;
-
-				// if we did not succeed at moving either point
-				if(x == testPoint.X && y == testPoint.Y)
-				{
-					x += 1;
-				}
-				polygonsToImageTransform.inverse_transform(ref x, ref y);
-				result = new IntPoint(Round(x), Round(y));
-			}
-
-			return movedPoint;
-		}
-
-		private bool CheckInsetPixel(int current, int xi, int ox, int yi, int oy, ref IntPoint result)
-		{
-			int value = GetInsetMapValue(xi + ox, yi + oy);
-			if (value > current)
-			{
-				result += new IntPoint(ox, oy);
-				return true;
-			}
-
-			return false;
-		}
-
-		private int GetInsetMapValue(int xi, int yi)
-		{
-			if (xi >= 0 && xi < InsetMap.Width
-				&& yi >= 0 && yi < InsetMap.Height)
-			{
-				var buffer = InsetMap.GetBuffer();
-				var offset = InsetMap.GetBufferOffsetXY(xi, yi);
-				return buffer[offset];
-			}
-
-			return 0;
-		}
-
-		public QTPolygonsExtensions.InsideState PointIsInside(IntPoint testPoint)
-		{
-			if (!usingPathingCache)
-			{
-				if (Polygons.PointIsInside(testPoint, EdgeQuadTrees, PointQuadTrees))
-				{
-					return QTPolygonsExtensions.InsideState.Inside;
-				}
-
-				return QTPolygonsExtensions.InsideState.Outside;
-			}
-
-			// translate the test point to the image coordinates
-			double xd = testPoint.X;
-			double yd = testPoint.Y;
-			polygonsToImageTransform.transform(ref xd, ref yd);
-			int xi = (int)Round(xd);
-			int yi = (int)Round(yd);
-
-			int pixelSum = 0;
-			for (int offsetX = -1; offsetX <= 1; offsetX++)
-			{
-				for (int offsetY = -1; offsetY <= 1; offsetY++)
-				{
-					int x = xi + offsetX;
-					int y = yi + offsetY;
-					if (x >= 0 && x < InsideCache.Width
-						&& y >= 0 && y < InsideCache.Height)
-					{
-						pixelSum += InsideCache.GetBuffer()[InsideCache.GetBufferOffsetXY(x, y)];
-					}
-				}
-			}
-
-			if (pixelSum == 0)
-			{
-				return QTPolygonsExtensions.InsideState.Outside;
-			}
-			else if (pixelSum / 9 == 255)
-			{
-				return QTPolygonsExtensions.InsideState.Inside;
-			}
-
-			// The cache could not definitively tell us, so check the polygons
-			if (Polygons.PointIsInside(testPoint, EdgeQuadTrees, PointQuadTrees))
-			{
-				return QTPolygonsExtensions.InsideState.Inside;
-			}
-
-			return QTPolygonsExtensions.InsideState.Outside;
-		}
-
-		private void GenerateIsideCache()
-		{
-			int width = (int)Round(polygonBounds.Width() / unitsPerPixel);
-			int height = (int)Round(polygonBounds.Height() / unitsPerPixel);
-
-			InsideCache = new ImageBuffer(width + 4, height + 4, 8, new blender_gray(1));
-
-			// Set the transform to image space
-			polygonsToImageTransform = Affine.NewIdentity();
-			// move it to 0, 0
-			polygonsToImageTransform *= Affine.NewTranslation(-polygonBounds.minX, -polygonBounds.minY);
-			// scale to fit cache
-			polygonsToImageTransform *= Affine.NewScaling(width / (double)polygonBounds.Width(), height / (double)polygonBounds.Height());
-			// and move it in 2 pixels
-			polygonsToImageTransform *= Affine.NewTranslation(2, 2);
-
-			// and render the polygon to the image
-			InsideCache.NewGraphics2D().Render(new VertexSourceApplyTransform(CreatePathStorage(Polygons), polygonsToImageTransform), Color.White);
-
-			// Now lets create an image that we can use to move points inside the outline
-			// First create an image that is fully set on all color values of the original image
-			InsetMap = new ImageBuffer(InsideCache);
-			InsetMap.DoThreshold(1);
-			// Then erode the image multiple times to get the a map of desired insets
-			int count = 8;
-			int step = 255/count;
-			ImageBuffer last = InsetMap;
-			for (int i = 0; i < count; i++)
-			{
-				var erode = new ImageBuffer(last);
-				erode.DoErode3x3Binary(255);
-				Paint(InsetMap, erode, (i + 1) * step);
-				last = erode;
-			}
-		}
-
-		private void Paint(ImageBuffer dest, ImageBuffer source, int level)
-		{
-			int height = source.Height;
-			int width = source.Width;
-			int sourceStrideInBytes = source.StrideInBytes();
-			int destStrideInBytes = dest.StrideInBytes();
-			byte[] sourceBuffer = source.GetBuffer();
-			byte[] destBuffer = dest.GetBuffer();
-
-			for (int y = 1; y < height - 1; y++)
-			{
-				int offset = source.GetBufferOffsetY(y);
-				for (int x = 1; x < width - 1; x++)
-				{
-					if (destBuffer[offset] == 255 // the dest is white
-						&& sourceBuffer[offset] == 0) // the dest is cleared
-					{
-						destBuffer[offset] = (byte)level;
-					}
-					offset++;
 				}
 			}
 		}
